Moved getNews to News.ts, manually create news item object, pool.query instead of create new connection, cache changed to HOUR
